requirements.txt
fastapi
pydantic
uvicorn
mysql-connector-python
python-dotenv
pandas
python-multipart
passlib[argon2]
pydantic[email]
email-validator
